Route: Avoid calculating a NaN travelPerTurn
If getOptimalSpeed() is called
when the car has already arrived at the ferry terminal,
distToTravel is 0.
It is divided by the time it would take to arrive (also 0),
which gives a travelPerTurn value of NaN.

Avoid that by just returning the default travelPerTurn.
The value should not matter at this point.

Also, remove the warning message printed
when the agent can move exactly
as much as required to reach the destination.
With Route adjusting travelPerTurn to do just that,
this is no longer an unusual condition.