DefaultAgent: send and receive times
Send the expected time to arrive at a ferry to the ferry terminal.
Receive the best time to arrive at a ferry from the ferry terminal.

Don't use the ferry terminal agent's methods
for getting the ferry departure time directly.

FerryTerminalAgent: send the time
just before the next ferry departure time
to the TravellingAgents that asked for it
by sending the expected arrival time.

Extract a TravellingAgent interface. It should be implemented by
all agents that communicate with FerryTerminalAgent.

TODO: Make the FerryAgent send the proper expected arrival time
to the FerryTerminalAgent.
TODO: Make the FerryTerminalAgent send the departure time
to the FerryAgent.

TODO: Make the FerryTerminalAgent create a dynamic schedule.